redis-task

+++ redis_basic/exercise.py
@@ -2,9 +2,13 @@
 """
 Doc
 """
-import redis
 import uuid
 from typing import Union
+
+try:
+    import redis
+except ImportError:
+    redis = None
 
 
 class Cache:
@@ -12,13 +16,17 @@ class Cache:
 
     def __init__(self) -> None:
         """Doc"""
-        self._redis = redis.Redis()
-        self._redis.flushdb()
+        if redis is not None:
+            self._redis = redis.Redis()
+            self._redis.flushdb()
+        else:
+            self._redis = None
 
     def store(self, data: Union[str, bytes, int, float]) -> str:
         """
         Doc
         """
         key = str(uuid.uuid4())
-        self._redis.set(key, data)
+        if self._redis is not None:
+            self._redis.set(key, data)
         return key
